requirements.txt
Flask==3.1.0
openai==1.60.1
twilio==9.4.3
flask-cors==5.0.0
flask_pymongo==2.3.0
python-dotenv==1.0.1
twilio==9.4.3
openai==1.60.1
